packages/flask/pyproject.toml
[project]
name = "korea-auth"
version = "0.1.0"
description = "Korean social login token verification for Flask/Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "pyjwt>=2.6.0",
    "cryptography>=39.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "responses>=0.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["korea_auth"]
